Added a filler for the case when the index can be held in the memory
